Revert recent ill-advised test case changes.
Commit 6d7b5ed1bd741250d5b672587323e62e0647633f cannot work on Windows,
because it lacks symlink support. While the bug fix in commit
90511536605c3d98e90e07b22729306345ed62a3 is correct as far as I know,
the test case changes depend on the previous commit, so this will
have to live without test coverage until we can come up with a better
solution. Commit 45bba563d6bb3de02fec83e15585d54a5556c6d4 was a test
case bug fix on top of those two, to prevent failures on Linux, so that
has to come out as well.

Per the buildfarm, CI, and Thomas Munro.